requirements.txt
Pillow
crcmod
distribute
pyserial
requests
wsgiref
